feat: moved log package to pkg
