+++ cmd/analyze.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/pkg/browser"
 	"github.com/qlik-oss/corectl/internal"
-	"github.com/qlik-oss/corectl/internal/log"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"github.com/qlik-oss/corectl/printer"
 	"github.com/spf13/cobra"
 )
