+++ cmd/bookmark.go
@@ -2,8 +2,8 @@ package cmd
 
 import (
 	"github.com/qlik-oss/corectl/internal"
-	"github.com/qlik-oss/corectl/internal/log"
 	"github.com/qlik-oss/corectl/pkg/boot"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"github.com/qlik-oss/corectl/printer"
 	"github.com/spf13/cobra"
 )
