+++ cmd/dimension.go
@@ -2,9 +2,9 @@ package cmd
 
 import (
 	"github.com/qlik-oss/corectl/internal"
-	"github.com/qlik-oss/corectl/internal/log"
 	"github.com/qlik-oss/corectl/pkg/boot"
 	"github.com/qlik-oss/corectl/pkg/dynconf"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"github.com/qlik-oss/corectl/printer"
 	"github.com/spf13/cobra"
 )
