+++ cmd/utils.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/google/go-github/v27/github"
 	ver "github.com/hashicorp/go-version"
-	"github.com/qlik-oss/corectl/internal/log"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"github.com/spf13/cobra"
 )
