+++ internal/measure.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 	"sort"
 
-	"github.com/qlik-oss/corectl/internal/log"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"github.com/qlik-oss/enigma-go"
 )
