+++ internal/modeldata.go
@@ -2,7 +2,7 @@ package internal
 
 import (
 	"context"
-	"github.com/qlik-oss/corectl/internal/log"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"github.com/qlik-oss/corectl/pkg/rest"
 	"github.com/qlik-oss/enigma-go"
 )
