+++ internal/systemtable.go
@@ -3,7 +3,7 @@ package internal
 import (
 	"context"
 
-	"github.com/qlik-oss/corectl/internal/log"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"github.com/qlik-oss/enigma-go"
 )
