+++ pkg/boot/knownapps.go
@@ -6,7 +6,7 @@ import (
 	"path"
 	"runtime"
 
-	"github.com/qlik-oss/corectl/internal/log"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"gopkg.in/yaml.v2"
 )
