+++ pkg/log/mode.go
