+++ printer/connections.go
@@ -4,7 +4,7 @@ import (
 	"os"
 
 	"github.com/olekukonko/tablewriter"
-	"github.com/qlik-oss/corectl/internal/log"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"github.com/qlik-oss/enigma-go"
 )
