+++ printer/entities.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/olekukonko/tablewriter"
 	"github.com/qlik-oss/corectl/internal"
-	"github.com/qlik-oss/corectl/internal/log"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"github.com/qlik-oss/enigma-go"
 )
