+++ test/clientcommand/main.go
@@ -2,9 +2,9 @@ package main
 
 import (
 	"fmt"
-	"github.com/qlik-oss/corectl/internal/log"
 	"github.com/qlik-oss/corectl/pkg/boot"
 	"github.com/qlik-oss/corectl/pkg/commands/standard"
+	"github.com/qlik-oss/corectl/pkg/log"
 	"github.com/spf13/cobra"
 	"strings"
 )
